move stringlib c and header files to more formal location in tree
