+++ lib/libac/cpu/simd_cpuid.c
@@ -51,7 +51,7 @@
 
 #include "aircrack-ng/cpu/simd_cpuid.h"
 #include "aircrack-ng/support/common.h"
-#include "aircrack-ng/string/string.h"
+#include "stringlib/string.h"
 
 #ifdef __linux__
 #define CPUFREQ_CPU0C "/sys/devices/system/cpu/cpu0/cpufreq/scaling_cur_freq"
